requirements.txt
# Kraken API and HTTP requests
requests>=2.31.0
krakenex>=2.2.0

# WebSocket support
websocket-client>=1.6.0

# Environment variables
python-dotenv>=1.0.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Logging
colorlog>=6.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
